requirements.txt
boto3==1.19.11
flask==2.0.2
Passlib==1.7.4
Flask-Caching==1.10.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask_Cors==3.0.10
Flask-Migrate==3.1.0
SQLAlchemy==1.4.26
requests==2.26.0
waitress==2.1.1
marshmallow-sqlalchemy==0.26.1
pytz==2021.3
PyMySQL==1.0.2
python-dotenv==0.19.0
python-slugify==5.0.2
Werkzeug==2.0.1
alembic==1.7.1
debugpy